Convert everything from sfa to sys2
sfa:   standard flow axis
sys2:  coordinate system 2

sfa only knew about the standard flow direction at each fov

sys2 is a full basis with two given basis vectors e1 e2.
To make a right handed system, e3 is implied but not currently
used.

The choice of e1 and e2 are in the config file.  e1 is supposed
to be chosen to be aligned with some standard flow direction.
In current tests, e1 is chosen to be the flow if the culture
is 2pi.u+.  Then e2 is chosen to point away from the culture
center.  This makes e3 point down.

In the future, sys3 will be implemented to point in the actual
global flow direction.  If we make sys3 e2 also point away
from the center of the culture, 2pi.u- will then make e3 point
upward.