Create sys3 basis in the fafovs
Note there has been a change in the way I input sys2 in
the config file since the last commit but without needing
a commit change.  Now sys2 is defined to have e2 pointing
to the right as you look along the flow.  This makes it
point inward.

This made no difference to any calculation since sys2_e2 was
not being used for anything.

But it's starting to be important as I work on sys3.  Now
sys3 will be the same as sys2 if the culture is truly 2pi.u+.
But sys3 will have a pi rotation relative to sys2 if the
culture is actually 2pi.u-.  This means that in sys3, e2
points to the center of the culture if 2pi.u+ and points
outward if 2pi.u-.

sys3 isn't used for anything yet.