Add fun.py module file from ipro 0115v10a.
This file was created and tested in 20220 ipro 0115v10a.
It has a bunch of accessory functions for doing vector
and matrix manipulation, and creating arrays useable
by pyplot.

It uses numpy arrays for vectors and matrices.

The plan is to convert much of asti-tracul code to use numpy
arrays, to use the fun.py accessory functions, and to alter
fun.py for this project.

In the future, ipro 0115v10a or some other ipro might be used
to turn fun.py stuff into its own library.  At that time, we
might start using that lib instead of a locally altered fun.py.