Add vid and n_track to table 2 output
Note that there is always a vid number here even though
that vid might be absent from traspe output because the
vid had no tracks to analyze and so no asti-trackpy output.

Exporting the vid is the c#.data file in the line with its
FOV is super handy into comparing traspe data to tracul
data.